remove plotter base class
