+++ waveform_editor/gui/plotter_view.py
@@ -1,16 +1,17 @@
 import holoviews as hv
+import panel as pn
 import param
+from panel.viewable import Viewer
 
-from waveform_editor.gui.plotter_base import PlotterBase
 
-
-class PlotterView(PlotterBase):
+class PlotterView(Viewer):
     """Class to plot multiple waveforms in view mode."""
 
     plotted_waveforms = param.Dict(default={})
 
     def __init__(self, **params):
         super().__init__(**params)
+        self.pane = pn.pane.HoloViews(sizing_mode="stretch_both")
         self.update_plot()
 
     @param.depends("plotted_waveforms", watch=True)
@@ -30,3 +31,26 @@ class PlotterView(PlotterBase):
 
         overlay = hv.Overlay(curves).opts(title="", show_legend=True)
         self.pane.object = overlay
+
+    def plot_waveform(self, waveform):
+        """
+        Store the tendencies of a waveform into a holoviews curve.
+
+        Args:
+            waveform: The waveform to convert to a holoviews curve.
+
+        Returns:
+            A Holoviews Curve object.
+        """
+        # TODO: The y axis should show the units of the plotted waveform
+        xlabel = "Time (s)"
+        ylabel = "Value"
+
+        if waveform is None or not waveform.tendencies:
+            return hv.Curve(([], []), xlabel, ylabel)
+        times, values = waveform.get_value()
+
+        return hv.Curve((times, values), xlabel, ylabel, label=waveform.name)
+
+    def __panel__(self):
+        return pn.Column(self.pane)
